64-bit safety fixes

+++ libnethack/src/do_wear.c
@@ -1,5 +1,5 @@
 /* vim:set cin ft=c sw=4 sts=4 ts=8 et ai cino=Ls\:0t0(0 : -*- mode:c;fill-column:80;tab-width:8;c-basic-offset:4;indent-tabs-mode:nil;c-file-style:"k&r" -*-*/
-/* Last modified by Alex Smith, 2013-12-22 */
+/* Last modified by Alex Smith, 2013-12-26 */
 /* Copyright (c) Stichting Mathematisch Centrum, Amsterdam, 1985. */
 /* NetHack may be freely redistributed.  See license for details. */
 
@@ -44,7 +44,7 @@ static const char * const c_slotnames_menu[] = {
     [os_quiver]  = "Quiver",
 };
 /* Hopefully, the compiler will optimize out the strlen of a literal string. */
-#define LONGEST_SLOTNAME strlen("Secondary weapon")
+#define LONGEST_SLOTNAME ((int)strlen("Secondary weapon"))
 
 static const char c_that_[] = "that";
 
@@ -1987,7 +1987,7 @@ doequip(const struct nh_cmd_arg *arg)
                     tos_first_equip + j];
                 sprintf(buf, "%*s %.*s%s", LONGEST_SLOTNAME + 5,
                         "->", BUFSZ - LONGEST_SLOTNAME - 7 -
-                        strlen(" (in progress)"),
+                        (int)strlen(" (in progress)"),
                         u.utracked[tos_first_equip + j] != &zeroobj ?
                         doname(u.utracked[tos_first_equip + j]) :
                         "(empty)",
